chore: separate file for contract interaction breakdown
