+++ .env.base
@@ -78,3 +78,6 @@ REACT_APP_MTPELERIN_REFERRAL_CODE=shapeshift
 
 # y.at
 REACT_APP_YAT_NODE_URL=https://a.y.at
+
+# etherscan
+REACT_APP_ETHERSCAN_API_KEY=XT8BI6VDYUGD9675X861ATHZNK3AN6HRMF
