+++ src/plugins/walletConnectToDapps/components/modal/callRequest/SendTransactionConfirmation.tsx
@@ -1,33 +1,19 @@
-import { CopyIcon } from '@chakra-ui/icons'
-import {
-  Box,
-  Button,
-  Divider,
-  HStack,
-  IconButton,
-  Image,
-  useColorModeValue,
-  VStack,
-} from '@chakra-ui/react'
+import { Box, Button, Image, useColorModeValue, VStack } from '@chakra-ui/react'
 import type {
   WalletConnectCallRequestResponseMap,
   WalletConnectEthSendTransactionCallRequest,
 } from '@shapeshiftoss/hdwallet-walletconnect-bridge/dist/types'
-import { CurrencyAmount } from '@uniswap/sdk'
-import _ from 'lodash'
-import { useContract } from 'plugins/walletConnectToDapps/ContractABIContext'
 import { useWalletConnect } from 'plugins/walletConnectToDapps/WalletConnectBridgeContext'
 import type { FC } from 'react'
-import { Fragment, useMemo } from 'react'
 import { FormProvider, useForm } from 'react-hook-form'
-import { FaCode, FaGasPump, FaWrench } from 'react-icons/fa'
+import { FaGasPump, FaWrench } from 'react-icons/fa'
 import { useTranslate } from 'react-polyglot'
 import { Card } from 'components/Card/Card'
 import { FoxIcon } from 'components/Icons/FoxIcon'
-import { MiddleEllipsis } from 'components/MiddleEllipsis/MiddleEllipsis'
-import { RawText, Text } from 'components/Text'
+import { Text } from 'components/Text'
 
 import { AddressSummaryCard } from './AddressSummaryCard'
+import { ContractInteractionBreakdown } from './ContractInteractionBreakdown'
 import { ModalSection } from './ModalSection'
 import { TransactionAdvancedParameters } from './TransactionAdvancedParameters'
 
@@ -44,18 +30,14 @@ export const SendTransactionConfirmation: FC<Props> = ({ request, onConfirm, onR
   const translate = useTranslate()
   const cardBg = useColorModeValue('white', 'gray.850')
 
-  const { contract } = useContract(request.to, request.chainId)
-  const transaction = useMemo(
-    () => contract?.parseTransaction({ data: request.data, value: request.value }),
-    [contract, request.data, request.value],
-  )
-
   const form = useForm<ConfirmData>({
     defaultValues: {
       nonce: request.nonce,
       gasLimit: request.gasLimit,
     },
   })
+
+  // const { fees } = useSendFees()
 
   const walletConnect = useWalletConnect()
   if (!walletConnect.bridge || !walletConnect.dapp) return null
@@ -104,61 +86,7 @@ export const SendTransactionConfirmation: FC<Props> = ({ request, onConfirm, onR
             mb={4}
           />
           <Card bg={cardBg} borderRadius='md' px={4} py={2}>
-            <ModalSection
-              title={
-                transaction?.name ??
-                translate(
-                  'plugins.walletConnectToDapps.modal.sendTransaction.contractInteraction.sendingEth',
-                )
-              }
-              icon={<FaCode />}
-            >
-              <Box pl={6} pt={2}>
-                <Text
-                  color='gray.500'
-                  fontWeight='medium'
-                  translation='plugins.walletConnectToDapps.modal.sendTransaction.contractInteraction.amount'
-                />
-                <RawText fontWeight='medium'>
-                  {/* TODO: what's the best way to format e.g. an ether amount with the appropriate amount of decimals? */}
-                  {CurrencyAmount.ether(request.value).toFixed()}
-                </RawText>
-                <Divider my={4} />
-                {!!transaction &&
-                  transaction.functionFragment.inputs.map((input, index) => (
-                    <Fragment key={index}>
-                      <RawText color='gray.500' fontWeight='medium'>
-                        {_.startCase(input.name)} ({input.type})
-                      </RawText>
-                      {input.type === 'bytes[]' ? (
-                        <MiddleEllipsis
-                          fontWeight='medium'
-                          value={transaction.args[index].toString()}
-                        />
-                      ) : (
-                        <RawText fontWeight='normal'>{transaction.args[index].toString()}</RawText>
-                      )}
-                      <Divider my={4} />
-                    </Fragment>
-                  ))}
-
-                <Text
-                  color='gray.500'
-                  fontWeight='medium'
-                  translation='plugins.walletConnectToDapps.modal.sendTransaction.contractInteraction.data'
-                />
-                <HStack>
-                  <MiddleEllipsis value={request.data} fontWeight='medium' />
-                  <IconButton
-                    size='small'
-                    variant='ghost'
-                    aria-label='Copy'
-                    icon={<CopyIcon />}
-                    onClick={() => navigator.clipboard.writeText(request.data)}
-                  />
-                </HStack>
-              </Box>
-            </ModalSection>
+            <ContractInteractionBreakdown request={request} />
           </Card>
         </Box>
 
@@ -167,7 +95,7 @@ export const SendTransactionConfirmation: FC<Props> = ({ request, onConfirm, onR
           icon={<FaGasPump />}
           defaultOpen={false}
         >
-          todo
+          {/* <TxFeeRadioGroup fees={fees} /> */}
         </ModalSection>
 
         <ModalSection
